# [DISCO-2473] refactor: Reuse HTTP client for accuweather backend