## References

JIRA: [DISCO-2473](https://mozilla-hub.atlassian.net/browse/DISCO-2473)
GitHub: N/A

## Description

Profiled Merino via Scalene and the result suggested the `AsyncClient` creation in the AccuWeather backend really requires a lot of CPU cycles and Merino created an AsyncClient for each request!

The benchmark results show a 5-6x boost in throughput and latency with the patch.

**Before: 225 QPS / 522 ms (p95 latency)
After: 1160 QPS / 115 ms (p95 latency)** 



## PR Review Checklist

_Put an `x` in the boxes that apply_

- [x] This PR conforms to the [Contribution Guidelines](https://github.com/mozilla-services/merino-py/blob/main/CONTRIBUTING.md)
- [x] The PR title starts with the JIRA issue reference, format `[DISCO-####]`, and has the same title (if applicable)
- [x] `[load test: (abort|warn)]` keywords are applied (if applicable)
- [ ] [Documentation](https://github.com/mozilla-services/merino-py/tree/main/docs) has been updated (if applicable)
- [x] [Functional and performance test](https://github.com/mozilla-services/merino-py/blob/main/docs/dev/testing.md) coverage has been expanded and maintained (if applicable)


[DISCO-2473]: https://mozilla-hub.atlassian.net/browse/DISCO-2473?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ